==> R/7_Mapping traits.Rmd <==
---
title: "7_Mapping traits"
author: "Mireia Valle"
date: "10/7/2019"
output: html_document
---

#------------------------------------------------------------------------------------- #
#-LOADING Required Libraries 
#------------------------------------------------------------------------------------- #
#libraries for spatial analysis
```{r setup}
require(foreign)
require(sp)
library(raster)       #Main raster library with nearly all functions used in this analysis
library(rgdal)        #Spatial library - most functions used from rgdal are for vectors (shapefiles)
require(maptools)
```
#for data management
```{r setup}
library (tibble)
library (tidyverse) 
library(dplyr)        #a data wrangling library
library (data.table)
library (reshape)
library (dbplyr)
```

#See WORKING DIRECTORY
```{r setup}
getwd()
```
#------------------------------------------------------------------------------------- #
# LOADING DATA OF SPECIES WITH LIFE TRAITS INFORMATION GOTTEN FROM FISHLIFE AND FISHBASE
#------------------------------------------------------------------------------------- #
```{r setup}

traits_FD<- read.csv ("/home/valle/github_annex/traits_FD_ok.csv") 
as.tibble (traits_FD)
colnames (traits_FD)
summary (traits_FD)

colnames(traits_FD)[colnames(traits_FD)=="X"] <- "am_sid"

#Key by "am_sid"
traits_keyed <- data.table(traits_FD, key = "am_sid") 

#remove not keyed dataframe
rm (traits_FD)
```
#------------------------------------------------------------------------------------- #
# LOADING AQUAMAPS DATA 
#------------------------------------------------------------------------------------- #
#loading hs table with information about the probability of occurrence of species
```{r setup}
hs <- read.csv('/home/valle/github_annex/Aquamaps/d2018/hcaf_species_native_ver0816c_fixed.csv', sep=',', header = TRUE)

head (hs)

colnames (hs)

#important column in this table : "am_sid" which we are going to use as key
hs_keyed <- data.table(hs,  key = "am_sid")
head (hs_keyed)
class(hs_keyed)
#remove not keyed dataframe
rm(hs)
```
#------------------------------------------------------------------------------------- #
#CREATING NEW JOINED TABLE USING am_sid KEY
#------------------------------------------------------------------------------------- #
```{r setup}
species_hs <- hs_keyed[traits_keyed] %>%
  setkey(NULL)

head (species_hs)

colnames (species_hs)

#remove both dataframes
rm(traits_keyed)
rm (hs_keyed)

#unique am_sid values
unique (species_hs$am_sid) %>% length()#12204

#which is the structure
str(species_hs)

table (species_hs$bodyshape)

#levels of those variables that are factors
table (species_hs$demerspelag)

table (species_hs$EnvTemp)

#save 
#write.csv (species_hs , file = "species_hs.csv")

#read 
#species_hs <- read.csv('/home/valle/github_annex/species_hs.csv')
```
#------------------------------------------------------------------------------------- #
# Creating a raster with LOICID value #
#------------------------------------------------------------------------------------- #
```{r setup}
basemap <- raster (ext= extent (c(-180,180, -90, 90)), res=0.5)
values (basemap) <- 1:length (basemap)
basemap

#class       : RasterLayer 
#dimensions  : 360, 720, 259200  (nrow, ncol, ncell)
#resolution  : 0.5, 0.5  (x, y)
#extent      : -180, 180, -90, 90  (xmin, xmax, ymin, ymax)
#coord. ref. : +proj=longlat +datum=WGS84 +ellps=WGS84 +towgs84=0,0,0 
#data source : in memory
#names       : layer 
#values      : 1, 259200  (min, max)

plot (basemap)
writeRaster(basemap, "basemap", format = "GTiff")
```
#------------------------------------------------------------------------------------- #
#MAPPING THE DISTRIBUTION OF ALL 12,204 SPECIES 
#------------------------------------------------------------------------------------- #
```{r setup}

all_spp <-  species_hs %>%
  group_by (loiczid)%>%
  summarize(n_spp = n())
  
head (all_spp)

unique (all_spp$loiczid) %>% length()#168700

unique (all_spp$n_spp) %>% length()#3696

allspecies_map <- subs (basemap, all_spp, by= "loiczid", which= "n_spp")

plot (allspecies_map)

writeRaster(allspecies_map, "allspecies_map", format = "GTiff")
```
#------------------------------------------------------------------------------------- #
#MAPPING THE DISTRIBUTION OF demerspelag TRAIT RICHNESS
#------------------------------------------------------------------------------------- #
```{r setup}
all_demerspelag <- species_hs %>%
  group_by (loiczid)%>%
  summarize(n_distinct(demerspelag))

summarin()

head (all_demerspelag)

all_demerspelag_map <- subs (basemap, all_demerspelag, by= "loiczid", which= "n_distinct(demerspelag)")
plot (all_demerspelag_map, col=colorRampPalette(c("red", "blue", "green"))(3))
legend("topright", legend = c("1", "2", "3"), fill = colorRampPalette(c("red", "blue", "green"))(3))

summary (species_hs$demerspelag)

writeRaster(all_demerspelag, "all_demerspelag", format = "GTiff")
```
#------------------------------------------------------------------------------------- #
#MAPPING THE DISTRIBUTION OF EnvTemp TRAIT RICHNESS
#------------------------------------------------------------------------------------- #
```{r setup}
all_EnvTemp <- species_hs %>%
  group_by (loiczid)%>%
  summarize(n_distinct(EnvTemp))

head (all_EnvTemp)

all_EnvTemp_map <- subs (basemap, all_EnvTemp, by= "loiczid", which= "n_distinct(EnvTemp)")
plot (all_EnvTemp_map, col=rainbow(6))
legend("topright", legend = c("6", "5", "4", "3", "2", "1"), fill = rev(rainbow(6)))

summary (species_hs$EnvTemp)
#boreal  deep-water       polar subtropical   temperate    tropical 
#     114230    37169656      614062     7511610     2258710    13392189

writeRaster(all_EnvTemp, "all_EnvTemp", format = "GTiff")
```
#------------------------------------------------------------------------------------- #
#MAPPING THE DISTRIBUTION OF Loo TRAIT RICHNESS
#------------------------------------------------------------------------------------- #
```{r setup}
#Summarize Loo
summary (species_hs$Loo)

#Create 4 different group intervals
species_hs$Loo_groups<-cut(species_hs$Loo, c(0,25,50,100,1500), labels=c("1", "2", "3", "4"))
species_hs$Loogroups<-cut(species_hs$Loo, c(0,25,50,100,1500), labels=c("small", "medium", "large", "very_large"))
head (species_hs)

summary (species_hs$Loo_groups)
#       1        2        3        4 
#27838666 19657395  8540381  5024015 
#0-25       25-50     50-100    100-1500

all_Loo <- species_hs %>%
  group_by (loiczid)%>%
  summarize(n_distinct(Loogroups))

head (all_Loo)

table (all_Loo$"n_distinct(Loogroups)")
#     1      2      3      4 
#  5158   1511  14574 147457

all_Loo_map <- subs (basemap, all_Loo, by= "loiczid", which= "n_distinct(Loogroups)")
plot (all_Loo_map, col=rainbow(4))

writeRaster(all_Loo, "all_Loo", format = "GTiff")
```  
#MAPPING SMALL FISH 0-25
```{r setup}
smallfish <- subset (species_hs, species_hs$Loo_groups == 1 )
unique (smallfish$am_sid) %>% length()#4025

all_smallfish <-  smallfish %>%
  group_by (loiczid)%>%
  summarize(n_spp = n())
  
head (all_smallfish)


allsmallfish_map <- subs (basemap, all_smallfish, by= "loiczid", which= "n_spp")

plot (allsmallfish_map)

summary (all_smallfish$n_spp)

writeRaster(allsmallfish_map, "allsmallfish_map", format = "GTiff")
```
#MAPPING MEDIUM FISH 25-50
```{r setup}
mediumfish <- subset (species_hs, species_hs$Loo_groups == 2 )
unique (mediumfish$am_sid) %>% length()#5348

all_mediumfish <-  mediumfish %>%
  group_by (loiczid)%>%
  summarize(n_spp = n())
  
head (all_mediumfish)

allmediumfish_map <- subs (basemap, all_mediumfish, by= "loiczid", which= "n_spp")

plot (allmediumfish_map)

summary (all_mediumfish$n_spp)

writeRaster(allmediumfish_map, "allmediumfish_map", format = "GTiff")
```
#MAPPING LARGE FISH 50-100
```{r setup}
largefish <- subset (species_hs, species_hs$Loo_groups == 3 )
unique (largefish$am_sid) %>% length()#1899

all_largefish <-  largefish %>%
  group_by (loiczid)%>%
  summarize(n_spp = n())
  
head (all_largefish)

alllargefish_map <- subs (basemap, all_largefish, by= "loiczid", which= "n_spp")

plot (alllargefish_map)

summary (all_largefish$n_spp)

writeRaster(alllargefish_map, "alllargefish_map", format = "GTiff")
```
#MAPPING VERY LARGE FISH 100-1500
```{r setup}
verylargefish <- subset (species_hs, species_hs$Loo_groups == 4 )
unique (verylargefish$am_sid) %>% length()#932

all_verylargefish <-  verylargefish %>%
  group_by (loiczid)%>%
  summarize(n_spp = n())
  
head (all_verylargefish)

allverylargefish_map <- subs (basemap, all_verylargefish, by= "loiczid", which= "n_spp")

plot (allverylargefish_map)

summary (all_verylargefish$n_spp)

writeRaster(allverylargefish_map, "allverylargefish_map", format = "GTiff")
```
#MAPPING PROPORTION

proportion_small <- (allsmallfish_map*100)/allspecies_map
plot (proportion_small)

proportion_medium <- (allmediumfish_map*100)/allspecies_map
plot (proportion_medium)

proportion_large <- (alllargefish_map*100)/allspecies_map
plot (proportion_large)

proportion_verylarge <- (allverylargefish_map*100)/allspecies_map
plot (proportion_verylarge)

proportion_all <- (allspecies_map*100)/allspecies_map
plot (proportion_all)

par(mfrow=c(2,5))

plot (allspecies_map,
      main="Distribution of all species",
      axes=FALSE)
plot (allsmallfish_map,
      main="Distribution of under 25 cm group")
plot (allmediumfish_map,
      main="Distribution of 25 to 50 cm group")
plot (alllargefish_map,
      main="Distribution of 50 to 100 cm group")
plot (allverylargefish_map,
      main="Distribution of 100 to 1500 cm group")

plot (proportion_all,
      col=terrain.colors(10),
      main="Proportion of all fish compared to all fish",
      axes=FALSE)
plot (proportion_small,
      col=terrain.colors(10),
      main="Proportion of under 25 cm fish compared to all fish",
      axes=FALSE)
plot (proportion_medium,
      col=terrain.colors(10),
      main="Proportion of 25 to 50 cm fish compared to all fish",
      axes=FALSE)
plot (proportion_large,
      col=terrain.colors(10),
      main="Proportion of 50 to 100 cm fish compared to all fish",
      axes=FALSE)
plot (proportion_verylarge,
      col=terrain.colors(10),
      main="Proportion of 100 to 1500 cm fish compared to all fish",
      axes=FALSE)
#------------------------------------------------------------------------------------- #
#MAPPING THE DISTRIBUTION OF BODYSHAPE TRAIT RICHNESS
#------------------------------------------------------------------------------------- #
#Summarize 
table (species_hs$bodyshape)
#       eel_like        elongated  fusiform_normal short_andor_deep            other 
#         4682026         27178670         13071362          9486940          1472631

sum(is.na(species_hs$bodyshape)) #5168828

asdf <- subset (species_hs, species_hs$loiczid=="11858")


all_bodyshape <- species_hs %>%
  group_by (loiczid)%>%
  summarize(n_tot = n(), 
  n_dist = n_distinct(bodyshape[!is.na(bodyshape)]))
  
all_bodyshape_map <- subs (basemap, all_bodyshape, by= "loiczid", which= "n_dist")
plot (all_bodyshape_map, col=rainbow(5))

sum (all_bodyshape$n_tot)
table (all_bodyshape$n_dist)

writeRaster(all_Loo, "all_Loo", format = "GTiff")

all_bodyshape_pct <- species_hs %>%
  group_by (loiczid)%>%
  summarize(n_tot = n(), 
  pct_eel = (sum (bodyshape=="eel_like")/n_tot),
  pct_elon = (sum (bodyshape=="elongated")/n_tot),
  pct_fusi = (sum (bodyshape=="fusiform_normal")/n_tot),
  pct_short = (sum (bodyshape=="short_andor_deep")/n_tot),
  pct_other = (100 - "pct_ell" - "pct_elon" - "pct_fusi" - "pct_short" - "pct-short"),
  n_dist = n_distinct(bodyshape))

#Error in 100 - "pct_ell" : non-numeric argument to binary operator

head (all_Loo)

table (all_Loo$"n_distinct(Loogroups)")
#     1      2      3      4 
#  5158   1511  14574 147457

all_Loo_map <- subs (basemap, all_Loo, by= "loiczid", which= "n_distinct(Loogroups)")
plot (all_Loo_map, col=rainbow(4))

writeRaster(all_Loo, "all_Loo", format = "GTiff")
